move power event handlers
