+++ modules/system/power/power_handler.py
@@ -1,0 +1,53 @@
+from system.eventbus import eventbus
+from system.power import events
+import power_event
+
+
+class PowerEventHandler:
+    def ChargeEventHandler(self):
+        eventbus.emit(
+            events.RequestChargeEvent(events.PowerEvent("Charge Cycle change"))
+        )
+
+    def FaultEventHandler(self):
+        eventbus.emit(events.RequestChargeFaultEvent(events.PowerEvent("Charge Fault")))
+
+    def HostAttachHandler(self):
+        eventbus.emit(
+            events.RequestHostAttachEvent(events.PowerEvent("Host attatched"))
+        )
+
+    def HostDetachHandler(self):
+        eventbus.emit(
+            events.RequestHostDetachEvent(events.PowerEvent("Host Detatched"))
+        )
+
+    def DeviceAttachHandler(self):
+        eventbus.emit(
+            events.RequestDeviceAttachEvent(events.PowerEvent("Device attatched"))
+        )
+
+    def DeviceDetachHandler(self):
+        eventbus.emit(
+            events.RequestDeviceDetachEvent(events.PowerEvent("Device detatched"))
+        )
+
+    def LanyardAttachHandler(self):
+        eventbus.emit(
+            events.RequestLanyardAttachEvent(events.PowerEvent("Lanyard attatched"))
+        )
+
+    def LanyardDetachHandler(self):
+        eventbus.emit(
+            events.RequestLanyardDetachEvent(events.PowerEvent("Lanyard Detatched"))
+        )
+
+
+power_event.set_charge_cb(PowerEventHandler.ChargeEventHandler)
+power_event.set_device_attach_cb(PowerEventHandler.DeviceAttachHandler)
+power_event.set_device_detach_cb(PowerEventHandler.DeviceDetachHandler)
+power_event.set_fault_cb(PowerEventHandler.FaultEventHandler)
+power_event.set_host_attach_cb(PowerEventHandler.HostAttachHandler)
+power_event.set_host_detach_cb(PowerEventHandler.HostDetachHandler)
+power_event.set_lanyard_attach_cb(PowerEventHandler.LanyardAttachHandler)
+power_event.set_lanyard_detach_cb(PowerEventHandler.LanyardDetachHandler)
